+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] 2025-09-05
+### Changed
+- Applies class `czcr-reading-empty` or `czcr-readings-has` to `body` based on
+base on the reading queue status.
+
 ## [1.4.1] 2025-08-30
 ### Changed
 - Update all colors to RGBa
@@ -54,7 +59,8 @@ page change.
 - Initial Commit
 
 
-[Unreleased]: https://github.com/erremauro/cz-continue-reading/compare/v1.4.1...HEAD
+[Unreleased]: https://github.com/erremauro/cz-continue-reading/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/erremauro/cz-continue-reading/releases/tag/v1.5.0
 [1.4.1]: https://github.com/erremauro/cz-continue-reading/releases/tag/v1.4.1
 [1.4.0]: https://github.com/erremauro/cz-continue-reading/releases/tag/v1.4.0
 [1.3.1]: https://github.com/erremauro/cz-continue-reading/releases/tag/v1.3.1